+++ king_phisher/security_keys.py
@@ -32,6 +32,7 @@
 
 import binascii
 import copy
+import hashlib
 import json
 import logging
 
@@ -39,10 +40,16 @@ from king_phisher import find
 from king_phisher import serializers
 from king_phisher import utilities
 
+import cryptography.hazmat.primitives.ciphers
+import cryptography.hazmat.primitives.ciphers.algorithms
+import cryptography.hazmat.primitives.ciphers.modes
+import cryptography.hazmat.primitives.padding as padding
+import cryptography.hazmat.backends as backends
 import ecdsa
 import ecdsa.curves
 import ecdsa.keys
 
+ciphers = cryptography.hazmat.primitives.ciphers
 ecdsa_curves = dict((c.name, c) for c in ecdsa.curves.curves)
 """
 A dictionary of :py:class:`ecdsa.curves.Curve` objects keyed by their
@@ -89,6 +96,73 @@ def _kwarg_curve(kwargs):
 	kwargs['curve'] = curve
 	return kwargs
 
+def openssl_decrypt_data(ciphertext, password, digest='sha256', encoding='utf-8'):
+	"""
+	Decrypt *ciphertext* in the same way as OpenSSL. For the meaning of
+	*digest* see the :py:func:`.openssl_derive_key_and_iv` function
+	documentation.
+
+	.. note::
+		This function can be used to decrypt ciphertext created with the
+		``openssl`` command line utility.
+
+		.. code-block:: none
+
+			openssl enc -e -aes-256-cbc -in file -out file.enc -md sha256
+
+	:param bytes ciphertext: The encrypted data to decrypt.
+	:param str password: The password to use when deriving the decryption key.
+	:param str digest: The name of hashing function to use to generate the key.
+	:param str encoding: The name of the encoding to use for the password.
+	:return: The decrypted data.
+	:rtype: bytes
+	"""
+	salt = b''
+	if ciphertext[:8] == b'Salted__':
+		salt = ciphertext[8:16]
+		ciphertext = ciphertext[16:]
+	my_key, my_iv = openssl_derive_key_and_iv(password, salt, 32, 16, digest=digest, encoding=encoding)
+
+	cipher = ciphers.Cipher(
+		ciphers.algorithms.AES(my_key),
+		ciphers.modes.CBC(my_iv),
+		backend=backends.default_backend()
+	)
+	decryptor = cipher.decryptor()
+	plaintext = decryptor.update(ciphertext) + decryptor.finalize()
+
+	unpadder = padding.PKCS7(cipher.algorithm.block_size).unpadder()
+	return unpadder.update(plaintext) + unpadder.finalize()
+
+def openssl_derive_key_and_iv(password, salt, key_length, iv_length, digest='sha256', encoding='utf-8'):
+	"""
+	Derive an encryption key and initialization vector (IV) in the same way as
+	OpenSSL.
+
+	.. note::
+		Different versions of OpenSSL use a different default value for the
+		*digest* function used to derive keys and initialization vectors. A
+		specific one can be used by passing the ``-md`` option to the
+		``openssl`` command.
+
+	:param str password: The password to use when deriving the key and IV.
+	:param bytes salt: A value to use as a salt for the operation.
+	:param int key_length: The length in bytes of the key to return.
+	:param int iv_length: The length in bytes of the IV to return.
+	:param str digest: The name of hashing function to use to generate the key.
+	:param str encoding: The name of the encoding to use for the password.
+	:return: The key and IV as a tuple.
+	:rtype: tuple
+	"""
+	password = password.encode(encoding)
+	digest_function = getattr(hashlib, digest)
+	chunk = b''
+	data = b''
+	while len(data) < key_length + iv_length:
+		chunk = digest_function(chunk + password + salt).digest()
+		data += chunk
+	return data[:key_length], data[key_length:key_length + iv_length]
+
 class SigningKey(ecdsa.SigningKey, object):
 	@classmethod
 	def from_secret_exponent(cls, *args, **kwargs):
@@ -104,7 +178,38 @@ class SigningKey(ecdsa.SigningKey, object):
 
 	@classmethod
 	def from_dict(cls, value, encoding='base64'):
+		"""
+		Load the signing key from the specified dict object.
+
+		:param dict value: The dictionary to load the key data from.
+		:param str encoding: The encoding of the required 'data' key.
+		:return: The new signing key.
+		:rtype: :py:class:`.SigningKey`
+		"""
 		return _key_cls_from_dict(cls, value, encoding=encoding)
+
+	@classmethod
+	def from_file(cls, file_path, password=None, encoding='utf-8'):
+		"""
+		Load the signing key from the specified file. If *password* is
+		specified, the file is assumed to have been encoded using OpenSSL using
+		``aes-256-cbc`` with ``sha256`` as the message digest.
+
+		:param str file_path: The path to the file to load.
+		:param str password: An optional password to use for decrypting the file.
+		:param str encoding: The encoding of the data.
+		:return: A tuple of the key's ID, and the new :py:class:`.SigningKey` instance.
+		:rtype: tuple
+		"""
+		with open(file_path, 'rb') as file_h:
+			file_data = file_h.read()
+		if password:
+			file_data = openssl_decrypt_data(file_data, password, encoding=encoding)
+
+		file_data = file_data.decode(encoding)
+		file_data = serializers.JSON.loads(file_data)
+		utilities.validate_json_schema(file_data, 'king-phisher.security.key')
+		return file_data['id'], cls.from_dict(file_data['signing-key'], encoding=file_data.pop('encoding', 'base64'))
 
 	def sign_dict(self, data, signature_encoding='base64'):
 		"""
@@ -116,9 +221,11 @@ class SigningKey(ecdsa.SigningKey, object):
 
 		:param dict data: The dictionary of data to sign.
 		:param str signature_encoding: The encoding name of the signature data.
+		:return: The dictionary object is returned with the 'signature' key added.
 		"""
 		utilities.assert_arg_type(data, dict, arg_pos=1)
 		data = copy.copy(data)
+		data.pop('signature', None)  # remove a pre-existing signature
 		json_data = json.dumps(data, sort_keys=True).encode('utf-8')
 		data['signature'] = _encoding_data(self.sign(json_data), encoding=signature_encoding)
 		return data
@@ -189,19 +296,16 @@ class SecurityKeys(object):
 			return 0
 		with open(file_path, 'r') as file_h:
 			key_store = serializers.JSON.load(file_h)
-		key_store = key_store.get('keys', [])
+		utilities.validate_json_schema(key_store, 'king-phisher.security')
+		key_store = key_store['keys']
 		loaded = 0
 		for key_idx, key in enumerate(key_store, 1):
-			identifier = key.get('id')
-			if not identifier:
-				self.logger.warning("skipping loading {0}:{1} due to missing id".format(file_name, key_idx))
-				continue
+			identifier = key['id']
 			if identifier in self.keys:
 				self.logger.warning("skipping loading {0}:{1} due to a duplicate id".format(file_name, key_idx))
 				continue
-			if 'verifying-key' in key:
-				verifying_key = key['verifying-key']
-				key['verifying-key'] = VerifyingKey.from_dict(verifying_key, encoding=verifying_key.pop('encoding', 'base64'))
+			verifying_key = key['verifying-key']
+			key['verifying-key'] = VerifyingKey.from_dict(verifying_key, encoding=verifying_key.pop('encoding', 'base64'))
 			self.keys[identifier] = key
 			self.logger.debug("loaded key id: {0} from: {1}".format(identifier, file_path))
 			loaded += 1