+++ king_phisher/server/server.py
@@ -169,11 +169,17 @@ class KingPhisherRequestHandler(advancedhttpserver.RequestHandler):
 			self.semaphore_acquire()
 		else:
 			self.adjust_path()
-		http_method_handler = getattr(super(KingPhisherRequestHandler, self), 'do_' + self.command)
+
+		http_method_handler = None
 		try:
+			signals.request_handle.send(self)
+			http_method_handler = getattr(super(KingPhisherRequestHandler, self), 'do_' + self.command)
 			http_method_handler(*args, **kwargs)
 		except errors.KingPhisherAbortRequestError as error:
-			self.logger.info('http request aborted')
+			if http_method_handler is None:
+				self.logger.debug('http request aborted by a signal handler')
+			else:
+				self.logger.info('http request aborted')
 			if not error.response_sent:
 				self.respond_not_found()
 		finally:
@@ -443,9 +449,7 @@ class KingPhisherRequestHandler(advancedhttpserver.RequestHandler):
 		signals.safe_send('response-sent', self.logger, self, code=code, message=message)
 
 	def respond_file(self, file_path, attachment=False, query=None):
-		self.semaphore_acquire()
 		self._respond_file_check_id()
-		self.semaphore_release()
 		file_path = os.path.abspath(file_path)
 		mime_type = self.guess_mime_type(file_path)
 		if attachment or (mime_type != 'text/html' and mime_type != 'text/plain'):
@@ -547,10 +551,15 @@ class KingPhisherRequestHandler(advancedhttpserver.RequestHandler):
 			return
 		if not self.config.get('server.require_id'):
 			return
+
+		self.semaphore_acquire()
 		if self.message_id == self.config.get('server.secret_id'):
+			self.semaphore_release()
+			self.server.logger.debug('request received with the correct secret id')
 			return
 		# a valid campaign_id requires a valid message_id
 		if not self.campaign_id:
+			self.semaphore_release()
 			self.server.logger.warning('denying request due to lack of a valid id')
 			raise errors.KingPhisherAbortRequestError()
 
@@ -559,21 +568,25 @@ class KingPhisherRequestHandler(advancedhttpserver.RequestHandler):
 		query = session.query(db_models.LandingPage)
 		query = query.filter_by(campaign_id=self.campaign_id, hostname=self.vhost)
 		if query.count() == 0:
-			self.server.logger.warning('denying request with not found due to invalid hostname')
 			session.close()
+			self.semaphore_release()
+			self.server.logger.warning('denying request with not found due to invalid hostname')
 			raise errors.KingPhisherAbortRequestError()
 		if campaign.has_expired:
-			self.server.logger.warning('denying request because the campaign has expired')
 			session.close()
+			self.semaphore_release()
+			self.server.logger.warning('denying request because the campaign has expired')
 			raise errors.KingPhisherAbortRequestError()
 		if campaign.reject_after_credentials and self.visit_id is None:
 			query = session.query(db_models.Credential)
 			query = query.filter_by(message_id=self.message_id)
 			if query.count():
-				self.server.logger.warning('denying request because credentials were already harvested')
 				session.close()
+				self.semaphore_release()
+				self.server.logger.warning('denying request because credentials were already harvested')
 				raise errors.KingPhisherAbortRequestError()
 		session.close()
+		self.semaphore_release()
 		return
 
 	def respond_not_found(self):
@@ -738,7 +751,7 @@ class KingPhisherRequestHandler(advancedhttpserver.RequestHandler):
 		self.logger.info("handling a page visit for campaign id: {0} from IP address: {1}".format(self.campaign_id, client_ip))
 		message = db_manager.get_row_by_id(session, db_models.Message, self.message_id)
 
-		if message.opened is None and self.config.get_if_exists('server.set_message_opened_on_visit', True):
+		if message.opened is None and self.config.get('server.set_message_opened_on_visit'):
 			message.opened = db_models.current_timestamp()
 			message.opener_ip = self.get_client_ip()
 			message.opener_user_agent = self.headers.get('user-agent', None)